css/css-flexbox/parsing/align-content-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Flexible Box Layout Module Level 1: parsing align-content with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-flexbox/#propdef-align-content">
<meta name="assert" content="align-content supports only the grammar 'flex-start | flex-end | center | space-between | space-around | stretch'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("align-content", "auto");

test_invalid_value("align-content", "stretch center"); // Accepted by Edge/WebKit.
</script>
</body>
</html>
